All user queries (batch 1) complete

+++ src/main/scala/com/dandxy/model/user/GameId.scala
@@ -2,10 +2,10 @@ package com.dandxy.model.user
 
 import doobie.util.Meta
 
-final case class GameId(id: Double) extends AnyVal
+final case class GameId(id: Int) extends AnyVal
 
 object GameId {
 
-  implicit val meta: Meta[GameId] = Meta[Double].imap(GameId(_))(_.id)
+  implicit val meta: Meta[GameId] = Meta[Int].imap(GameId(_))(_.id)
 
 }
